Improve downloading of off-file
Change shape object so it expects the data instead of the off-file. This gets
rid of the ugly construction with the this_ variable. Also allow to only use the
local path as parameter to draw_shape instead of the whole URL. Also get rid of
dependency to Googles javascript for downloading the off-file.