Change direction light
Wasn entirely satisfied with the direction of the light. For the
12.S4xI.D1xI.mu5.off if needed more shadows (compared to the picture I
have). It seemed that the light came too much from the front. Now it
comes a bit more from the back.

Another problem was that there are some faces that share an edge that
has 45 degrees with the x-axis in the 2D image. These faces get the same
colour. In that case it is not good that the directional light has the
same value for x, and y. Now that light comes more from the top.

Note that the show-off script copies this light automatically to the
opposite direction as well (since if cos_a < 0, the opposite is taken).
This is done, because some faces for uniform polyhedra have two sides
and then it is impossible to decide in what direction the normal is
pointing. As a consequence it is unwise to make the ligth completely
come from the side (or the top)