Fix handling of faces that are edges/vertices
Off files generated by Antiprism can have faces with 2 vertices or even
one vertex. This makes sense to be able to draw the vertices as small
spheres and the edges as cylindres. The script seemed to handle this,
but is was buggy, since the faces array was initialised with the
specified size, so that last part of the faces array was undefined.